requirements.txt
fastapi
uvicorn
openai
influxdb
pyyaml
schedule